+++ MoreShipUpgrades/Managers/UpgradeBus.cs
@@ -1,5 +1,6 @@
 ﻿using GameNetcodeStuff;
 using HarmonyLib;
+using LethalLib.Extras;
 using LethalLib.Modules;
 using MoreShipUpgrades.Misc;
 using MoreShipUpgrades.Misc.TerminalNodes;
@@ -37,6 +38,8 @@ namespace MoreShipUpgrades.Managers
         PlayerControllerB localPlayer;
         HangarShipDoor hangarDoors;
 
+        internal Dictionary<string, SpawnableMapObjectDef> spawnableMapObjects = new Dictionary<string, SpawnableMapObjectDef>();
+        internal Dictionary<string, int> spawnableMapObjectsAmount = new Dictionary<string, int>();
         internal List<CustomTerminalNode> terminalNodes = new List<CustomTerminalNode>();
         internal Dictionary<string, GameObject> UpgradeObjects = new Dictionary<string, GameObject>();
         internal Dictionary<string, Item> ItemsToSync = new Dictionary<string, Item>();
@@ -355,6 +358,7 @@ namespace MoreShipUpgrades.Managers
             SetupQuantumDisruptorTerminalNode();
             SetupLethalDealsTerminalNode();
             SetupFasterDropPodTerminalNode();
+            SetupChargingBoosterTerminalNode();
             terminalNodes.Sort();
         }
         void SetupMarketInfluenceTerminalNode()
@@ -544,6 +548,14 @@ namespace MoreShipUpgrades.Managers
                                     true,
                                     PluginConfiguration.FASTER_DROP_POD_ENABLED.Value,
                                     PluginConfiguration.FASTER_DROP_POD_PRICE.Value);
+        }
+        void SetupChargingBoosterTerminalNode()
+        {
+            SetupMultiplePurchasableTerminalNode(ChargingBooster.UPGRADE_NAME,
+                                                true,
+                                                PluginConfiguration.CHARGING_BOOSTER_ENABLED.Value,
+                                                PluginConfiguration.CHARGING_BOOSTER_PRICE.Value,
+                                                ParseUpgradePrices(PluginConfiguration.CHARGING_BOOSTER_PRICES.Value));
         }
         /// <summary>
         /// Generic function where it adds a terminal node for an upgrade that can be purchased multiple times