+++ MoreShipUpgrades/Misc/PluginConfig.cs
@@ -54,7 +54,7 @@ namespace MoreShipUpgrades.Misc
         readonly ConfigFile configFile;
 
         // enabled disabled
-
+        public ConfigEntry<bool> CHARGING_BOOSTER_ENABLED {  get; set; }
         public ConfigEntry<bool> MARKET_INFLUENCE_ENABLED { get; set; }
         public ConfigEntry<bool> BARGAIN_CONNECTIONS_ENABLED { get; set; }
         public ConfigEntry<bool> LETHAL_DEALS_ENABLED { get; set; }
@@ -100,6 +100,7 @@ namespace MoreShipUpgrades.Misc
         public ConfigEntry<bool> LOCKSMITH_INDIVIDUAL { get; set; }
 
         // prices
+        public ConfigEntry<int> CHARGING_BOOSTER_PRICE { get; set; }
         public ConfigEntry<int> MARKET_INFLUENCE_PRICE { get; set; }
         public ConfigEntry<int> BARGAIN_CONNECTIONS_PRICE {  get; set; }
         public ConfigEntry<int> LETHAL_DEALS_PRICE { get; set; }
@@ -133,6 +134,10 @@ namespace MoreShipUpgrades.Misc
         public ConfigEntry<int> FASTER_DROP_POD_PRICE { get; set; }
 
         // attributes
+        public ConfigEntry<string> CHARGING_BOOSTER_PRICES {  get; set; }
+        public ConfigEntry<float> CHARGING_BOOSTER_COOLDOWN {  get; set; }
+        public ConfigEntry<float> CHARGING_BOOSTER_INCREMENTAL_COOLDOWN_DECREASE { get; set; }
+        public ConfigEntry<int> CHARGING_BOOSTER_CHARGE_PERCENTAGE { get; set; }
         public ConfigEntry<float> BIGGER_LUNGS_STAMINA_REGEN_INCREASE { get; set; }
         public ConfigEntry<float> BIGGER_LUNGS_JUMP_STAMINA_COST_DECREASE { get; set; }
         public ConfigEntry<int> PROTEIN_INCREMENT { get; set; }
@@ -300,6 +305,7 @@ namespace MoreShipUpgrades.Misc
         public ConfigEntry<int> BARGAIN_CONNECTIONS_INCREMENTAL_ITEM_AMOUNT { get; set; }
         public ConfigEntry<string> BARGAIN_CONNECTIONS_PRICES { get; set; }
         public ConfigEntry<float> FASTER_DROP_POD_TIMER { get; set; }
+        public ConfigEntry<int> EXTRACTION_CONTRACT_AMOUNT_MEDKITS {  get; set; }
 
 
         public PluginConfig(ConfigFile PluginConfiguration)
@@ -329,6 +335,7 @@ namespace MoreShipUpgrades.Misc
             CONTRACT_EXTRACT_WEIGHT = ConfigEntry(topSection,"Weight of an extraction human", 2.5f, "Subtract 1 and multiply by 100 (2.5 = 150lbs).");
             SCAV_VOLUME = ConfigEntry(topSection,"Volume of the scavenger voice clips", 0.25f, "0.0 - 1.0");
             MAIN_OBJECT_FURTHEST = ConfigEntry(topSection, "Spawn main object far away", true, "If true the main object for contracts will try spawn as far away from the main entrance as possible. If false it will spawn at a random location.");
+            EXTRACTION_CONTRACT_AMOUNT_MEDKITS = ConfigEntry(topSection, "Amount of medkits that can spawn in the Extraction contract", 3);
 
             // this is kind of dumb and I'd like to just use a comma seperated ConfigEntry<string> but this is much more foolproof
             DATA_CONTRACT = ConfigEntry(topSection,"Enable the data contract", true, "Make this false if you don't want the data contract");
@@ -336,6 +343,14 @@ namespace MoreShipUpgrades.Misc
             EXORCISM_CONTRACT = ConfigEntry(topSection,"Enable the exorcism contract", true, "Make this false if you don't want the exorcism contract");
             DEFUSAL_CONTRACT = ConfigEntry(topSection,"Enable the defusal contract", true, "Make this false if you don't want the defusal contract");
             EXTERMINATOR_CONTRACT = ConfigEntry(topSection,"Enable the exterminator contract", true, "Make this false if you don't want the exterminator contract");
+
+            topSection = ChargingBooster.UPGRADE_NAME;
+            CHARGING_BOOSTER_ENABLED = ConfigEntry(topSection, "Enable Charging Booster Upgrade", true, "Tier upgrade which allows charging items in a radar booster");
+            CHARGING_BOOSTER_PRICE = ConfigEntry(topSection, "Price of Charging Booster Upgrade", 300);
+            CHARGING_BOOSTER_PRICES = ConfigEntry(topSection, BaseUpgrade.PRICES_SECTION, "250,300,400", BaseUpgrade.PRICES_DESCRIPTION);
+            CHARGING_BOOSTER_COOLDOWN = ConfigEntry(topSection, "Cooldown of the charging station in radar boosters when used", 90f);
+            CHARGING_BOOSTER_INCREMENTAL_COOLDOWN_DECREASE = ConfigEntry(topSection, "Incremental cooldown decrease whenever the level of the upgrade is incremented", 10f);
+            CHARGING_BOOSTER_CHARGE_PERCENTAGE = ConfigEntry(topSection, "Percentage of charge that is added to the held item when using the radar booster charge station", 50);
             
             topSection = "_Misc_";
             SHARED_UPGRADES = ConfigEntry(topSection, "Convert all upgrades to be shared.", true, "If true this will ignore the individual shared upgrade option for all other upgrades and set all upgrades to be shared.");