+++ MoreShipUpgrades/UpgradeComponents/Contracts/ExtractionContract.cs
@@ -1,4 +1,5 @@
-﻿using MoreShipUpgrades.UpgradeComponents.Items.Contracts.BombDefusal;
+﻿using MoreShipUpgrades.UpgradeComponents.Items;
+using MoreShipUpgrades.UpgradeComponents.Items.Contracts.BombDefusal;
 using MoreShipUpgrades.UpgradeComponents.Items.Contracts.Extraction;
 using System;
 using System.Collections.Generic;
@@ -8,10 +9,10 @@ namespace MoreShipUpgrades.UpgradeComponents.Contracts
 {
     internal class ExtractionContract : ContractObject
     {
-        public override void Start()
+        public override void Start()    
         {
             contractType = "extraction";
-            if (GetComponent<PhysicsProp>().isInShipRoom || GetComponent<PhysicsProp>().isInElevator) // Already in the ship, don't make it disappear
+            if (GetComponent<Medkit>() == null && GetComponent<GrabbableObject>().isInShipRoom || GetComponent<GrabbableObject>().isInElevator) // Already in the ship, don't make it disappear
             {
                 ExtractPlayerScript scavenger = GetComponent<ExtractPlayerScript>();
                 scavenger.MakeScavengerGrabbable();