+++ ChangeLog.md
@@ -1,3 +1,17 @@
+## V 3.3.0 - 2024-??-??
+### Additions
+- Added "Charging Booster" upgrade which allows radar boosters to charge up player's held item by looking at the base of the booster. Has a cooldown after use which can be decremented by increasing the upgrade's level.
+  - Configurable charge percentage on use.
+  - Configurable cooldown time on use.
+  - Configurable incremental cooldown decrease on upgrade levelling. 
+- Added "Drop Pod Thrusters" upgrade which decrease the amount of time you need to wait for the store's drop pod to arrive on the moon.
+  - Configurable timer for the drop pod to arrive 
+- Added configuration to the amount of medkits that can spawn in the extraction contract
+- Added interaction with doors to lockpick them to trigger the "Locksmith" upgrade's minigame. This way, you don't have to bump into doors to lockpick them.
+### Fixes
+- Fixed medkit breaking when Stimpack upgrade is disabled
+
+
 ## V 3.2.5 - 2024-02-26
 - Fixed Sick Beats icon being a white square instead of a boombox
 - Fixed a couple of the items displaying a white square when held